// server/data/interfaces/prisonApi/PrisonerSentenceDetails.ts
export default interface PrisonerSentenceDetails {
  bookingId: number
  offenderNo: string
  firstName: string
  lastName: string
  agencyLocationId: string
  mostRecentActiveBooking: boolean
  sentenceDetail: {
    bookingId?: number
    actualParoleDate?: string
    automaticReleaseOverrideDate?: string
    automaticReleaseDate?: string
    conditionalReleaseOverrideDate?: string
    conditionalReleaseDate?: string
    confirmedReleaseDate?: string
    dtoPostRecallReleaseDateOverride?: string
    dtoPostRecallReleaseDate?: string
    detentionPostRecallReleaseDate?: string
    earlyTermDate?: string
    earlyRemovalSchemeEligibilityDate?: string
    etdCalculatedDate?: string
    etdOverrideDate?: string
    effectiveSentenceEndDate?: string
    homeDetentionCurfewActualDate?: string
    homeDetentionCurfewEligibilityDate?: string
    licenceExpiryCalculatedDate?: string
    licenceExpiryDate?: string
    lateTermDate?: string
    ltdOverrideDate?: string
    ltdCalculatedDate?: string
    mtdCalculatedDate?: string
    mtdOverrideDate?: string
    midTermDate?: string
    nonDtoReleaseDateType?: 'ARD' | 'CRD' | 'NPD' | 'PRRD'
    nonDtoReleaseDate?: string
    nonParoleOverrideDate?: string
    nonParoleDate?: string
    paroleEligibilityCalculatedDate?: string
    paroleEligibilityDate?: string
    paroleEligibilityOverrideDate?: string
    postRecallReleaseDate?: string
    postRecallReleaseOverrideDate?: string
    releaseDate?: string
    releaseOnTemporaryLicenceDate?: string
    sentenceExpiryCalculatedDate?: string
    sentenceExpiryDate?: string
    sentenceStartDate?: string
    tariffEarlyRemovalSchemeEligibilityDate?: string
    tariffDate?: string
    topupSupervisionExpiryDate?: string
    topupSupervisionExpiryOverrideDate?: string
  }
  dateOfBirth: string
  agencyLocationDesc: string
  internalLocationDesc: string
  facialImageId: number
}